Move all styling functions of author network graph to hook
