+++ src/components/Visualizations/Graphs/NetworkGraph.tsx
@@ -31,8 +31,6 @@ const radius = width / 10;
 
 const numberOfLabelsToShow = 100;
 
-const noGroupColor = '#a6a6a6';
-
 export const NetworkGraph = ({
   root,
   link_data,
@@ -42,59 +40,11 @@ export const NetworkGraph = ({
 }: INetworkGraphProps): ReactElement => {
   const [selectedNode, setSelectedNode] = useState<IADSApiAuthorNetworkNode>();
 
-  const { partition, arc, color, fontSize, line, citationLimit, readLimit, linkScale, labelTransform, textAnchor } =
+  const { partition, arc, fontSize, line, labelTransform, textAnchor, nodeFill, labelDisplay, strokeWidth } =
     useNetworkGraph(root, link_data, keyToUseAsValue, radius, numberOfLabelsToShow);
 
   // actaul tree root data for graph
   const graphRoot = useMemo(() => partition(root), [partition, root]);
-
-  // get color of node
-  const nodeFill = (d: NetworkHierarchyNode<IADSApiAuthorNetworkNode>) => {
-    if (d.depth === 0) {
-      return 'white';
-    }
-    if (d.depth === 1) {
-      const index = d.parent.children.indexOf(d);
-      if (index < 7) {
-        d.color = color(index.toString());
-        return d.color;
-      }
-      return noGroupColor;
-    }
-    if (d.depth === 2) {
-      // child nodes
-      if (!d.parent.color) {
-        return noGroupColor;
-      }
-      return d.parent.color;
-    }
-  };
-
-  // get the label's display setting based on type of view
-  const labelDisplay = (d: NetworkHierarchyNode<IADSApiAuthorNetworkNode>, key: string) => {
-    if (key == 'size') {
-      return 'block';
-    }
-    if (key == 'citation_count' && d.data.citation_count > citationLimit) {
-      return 'block';
-    }
-    if (key == 'read_count' && d.data.read_count > readLimit) {
-      return 'block';
-    }
-    return 'none';
-  };
-
-  // get link stroke width
-  const strokeWidth = (d: ILink, links: ILink[]) => {
-    // get link weight
-    const weight = links.filter((l) => {
-      return (
-        (l.source.data.name === d.source.data.name && l.target.data.name === d.target.data.name) ||
-        (l.target.data.name === d.source.data.name && l.source.data.name === d.target.data.name)
-      );
-    })[0].weight;
-    return linkScale(weight);
-  };
 
   // node linking data
   const getLinks = (linkData: number[][]) => {
